+++ CHANGELOG.md
@@ -1,2 +1,17 @@
+CHANGELOG
+=========
+
+* version actuelle (non versionné pour le moment) :
+  * Klipper : 
+    * Configuration du /etc/default/klipper
+    * Ajout du fichier "variables" pour le stockage 
+    * Installation et configuration du plugins lors du 1er boot
+  * Octodash :
+    * Installation des pré-requis + octodash
+    * Configuration de l'API KEY et du thème
+    * Configuration des interpréteurs : xinit + bash + zsh
+    * Autologin
+
+  
 * 2023031600 :
   * Réécriture des scripts from scratch